+++ ProducaoAPI/ProducaoAPI/Responses/MateriaPrimaResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace ProducaoAPI.Responses;
 
-public record MateriaPrimaResponse(int Id, string Nome, string Fornecedor, string Unidade, double Preco, bool Ativo);
+public record MateriaPrimaResponse(int Id, string Nome, string Fornecedor, string Unidade, decimal Preco, bool Ativo);
